+++ src/main/resources/log4j.properties
@@ -4,7 +4,7 @@ log4j.appender.console.target=System.err
 log4j.appender.console.layout=org.apache.log4j.PatternLayout
 log4j.appender.console.layout.ConversionPattern=%d{yy/MM/dd HH:mm:ss} %p %c{1}: %m%n
 
-log4j.logger.org.krapsh=DEBUG
+log4j.logger.org.karps=DEBUG
 log4j.logger.org.eclipse.jetty.util.component.AbstractLifeCycle=ERROR
 log4j.logger.org.apache.spark=INFO
 log4j.logger.org.apache.spark.repl.SparkIMain$exprTyper=INFO
