+++ src/main/scala/org/karps/ops/ColumnTransforms.scala
@@ -1,4 +1,4 @@
-package org.krapsh.ops
+package org.karps.ops
 
 
 import com.typesafe.scalalogging.slf4j.{StrictLogging => Logging}
@@ -6,9 +6,9 @@ import org.apache.spark.sql.functions._
 import org.apache.spark.sql.types._
 import org.apache.spark.sql.{Column, DataFrame, KrapshStubs}
 import org.apache.spark.sql.catalyst.analysis.FunctionRegistry
-import org.krapsh.{ColumnWithType, DataFrameWithType, KrapshException}
-import org.krapsh.ops.Extraction.{FieldName, FieldPath}
-import org.krapsh.structures._
+import org.karps.{ColumnWithType, DataFrameWithType, KrapshException}
+import org.karps.ops.Extraction.{FieldName, FieldPath}
+import org.karps.structures._
 import spray.json.{JsArray, JsObject, JsValue}
 
 import scala.util.{Failure, Success, Try}
@@ -17,7 +17,7 @@ import scala.util.{Failure, Success, Try}
 // TODO: refactor to use ColumnWithType, it will simplify things.
 object ColumnTransforms extends Logging {
 
-  import org.krapsh.structures.JsonSparkConversions.{getString, get, sequence}
+  import org.karps.structures.JsonSparkConversions.{getString, get, sequence}
 
   /**
    * Starts from an unrectified dataframe represented as a column, and
