+++ src/main/scala/org/karps/structures/spark_data.scala
@@ -7,17 +7,17 @@
 //  - the basic unit of type in Spark is the Row with a StructType, while the basic
 //    unit of data in Krapsh is the Cell with a DataType. The functions below make sure
 //    that the data can be converted back and forth.
-package org.krapsh.structures
+package org.karps.structures
 
 import spray.json.{JsArray, JsBoolean, JsNull, JsNumber, JsObject, JsString, JsValue, RootJsonFormat, RootJsonWriter}
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.catalyst.expressions.GenericRowWithSchema
 import org.apache.spark.sql.types._
-import org.krapsh.row.{AlgebraicRow, Cell, RowArray, RowCell}
+import org.karps.row.{AlgebraicRow, Cell, RowArray, RowCell}
 import spray.json.DefaultJsonProtocol._
 import spray.json._
 import DefaultJsonProtocol._
-import org.krapsh.KrapshException
+import org.karps.KrapshException
 
 import scala.util.{Failure, Success, Try}
 
@@ -425,10 +425,10 @@ object LocalSparkConversion {
 }
 
 /**
- * Performs a number of conversions between the data as represented by krapsh
+ * Performs a number of conversions between the data as represented by karps
  * and the data accepted by Spark.
  *
- * All the data in a dataframe or in a row has to be a struct, but krapsh also
+ * All the data in a dataframe or in a row has to be a struct, but karps also
  * accepts primitive, non-nullable datatypes to be represented in a dataframe
  * or in a row. This is why there is a normalization process.
  * All data which is a non-nullable, primitive type associated with column name "_1
